Move out Inventory CTA validations into its own module
We will reuse it for other inventory validations

[#96139208]
