+++ lib/g5_integrations_validations.rb
@@ -16,6 +16,7 @@ require "g5_integrations_validations/yardi_voyager_inventory_cta_url_validations
 require "g5_integrations_validations/yardi_voyager_inventory_rate_source_validations"
 require "g5_integrations_validations/yardi_voyager_inventory_rate_to_display_validations"
 require "g5_integrations_validations/real_page_inventory_validations"
+require "g5_integrations_validations/inventory_cta_validations"
 
 module G5IntegrationsValidations
 end
